+++ src/constants/index.ts
@@ -1,3 +1,0 @@
-// Temporary structure, should be ENV service later on
-export const API_URL = 'https://api.themoviedb.org/3';
-export const BEARER_TOKEN = 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJjZDI5MGEyNTczNTQ5OWJhMDIyM2VlZmMyYzlkMDhiNCIsIm5iZiI6MTc0MTExMDYwMS4xMjQsInN1YiI6IjY3YzczZDQ5ODRkYTIxMTMzZWFjOTBjMCIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.5DZnyBAnKXQIk3UkPA693pM0tlsFynffJ6yJwy5OYqw';